refactor: 좋아요 수정 (Form)
