+++ src/main/java/com/pickx3/domain/entity/portfolio_package/FavoritesForm.java
@@ -1,0 +1,16 @@
+package com.pickx3.domain.entity.portfolio_package;
+
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@Setter
+@Getter
+@NoArgsConstructor
+public class FavoritesForm {
+
+    private Long userNum;
+
+    private Long portfolioNum;
+
+}
