+++ src/main/java/com/pickx3/service/FavoritesService.java
@@ -1,13 +1,13 @@
 package com.pickx3.service;
 
-import com.pickx3.domain.entity.Favorites;
+import com.pickx3.domain.entity.portfolio_package.Favorites;
+import com.pickx3.domain.entity.portfolio_package.FavoritesForm;
 import com.pickx3.domain.entity.portfolio_package.Portfolio;
+import com.pickx3.domain.entity.portfolio_package.dto.PortfolioResponseDto;
 import com.pickx3.domain.entity.user_package.User;
 import com.pickx3.domain.repository.FavoriteRepository;
 import com.pickx3.domain.repository.PortfolioRepository;
 import com.pickx3.domain.repository.UserRepository;
-import com.pickx3.dto.FavoritesDto;
-import com.pickx3.domain.entity.portfolio_package.dto.PortfolioResponseDto;
 import com.querydsl.jpa.impl.JPAQueryFactory;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -36,9 +36,9 @@ public class FavoritesService {
     private EntityManager em;
 
     // 좋아요
-    public Long addLike(FavoritesDto favoritesDto){
-        Portfolio portfolio = portfolioRepository.findById(favoritesDto.getPortfolio().getId()).orElseThrow(() -> new IllegalArgumentException("id가 존재하지 않습니다." ));
-        User user = userRepository.findById(favoritesDto.getUser().getId()).orElseThrow(() -> new IllegalArgumentException("id가 존재하지 않습니다." ));
+    public Long addLike(FavoritesForm favoritesForm){
+        Portfolio portfolio = portfolioRepository.findById(favoritesForm.getPortfolioNum()).orElseThrow(() -> new IllegalArgumentException("id가 존재하지 않습니다." ));
+        User user = userRepository.findById(favoritesForm.getUserNum()).orElseThrow(() -> new IllegalArgumentException("id가 존재하지 않습니다." ));
 
         // 좋아요 유무 확인 .isPresent ( == ) 값이 있으면 True 반환
         if(favoriteRepository.findByUserAndPortfolio(user, portfolio).isPresent()){throw new RuntimeException();}
@@ -54,9 +54,9 @@ public class FavoritesService {
     }
 
     // 좋아요 취소
-    public void cancelLike(FavoritesDto favoritesDto) {
-        Portfolio portfolio = portfolioRepository.findById(favoritesDto.getPortfolio().getId()).orElseThrow(() -> new IllegalArgumentException("id가 존재하지 않습니다." ));
-        User user = userRepository.findById(favoritesDto.getUser().getId()).orElseThrow(() -> new IllegalArgumentException("id가 존재하지 않습니다." ));
+    public void cancelLike(FavoritesForm favoritesForm) {
+        Portfolio portfolio = portfolioRepository.findById(favoritesForm.getPortfolioNum()).orElseThrow(() -> new IllegalArgumentException("id가 존재하지 않습니다." ));
+        User user = userRepository.findById(favoritesForm.getUserNum()).orElseThrow(() -> new IllegalArgumentException("id가 존재하지 않습니다." ));
         Favorites favorites = favoriteRepository.findByUserAndPortfolio(user, portfolio).orElseThrow(() -> new IllegalArgumentException("id(user, portfolio)가 존재하지 않습니다." ));
 
         favoriteRepository.delete(favorites);
